+++ src/titiler/application/tests/routes/test_stac.py
@@ -182,12 +182,12 @@ def test_part(httpx, rio, app):
 
     # Missing Assets or Expression
     response = app.get(
-        "/stac/crop/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json"
+        "/stac/bbox/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json"
     )
     assert response.status_code == 400
 
     response = app.get(
-        "/stac/crop/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&assets=B01&rescale=0,1000&max_size=64"
+        "/stac/bbox/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&assets=B01&rescale=0,1000&max_size=64"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
@@ -196,7 +196,7 @@ def test_part(httpx, rio, app):
     assert meta["height"] == 14
 
     response = app.get(
-        "/stac/crop/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&assets=B01&rescale=0,1000&max_size=64&width=128&height=128"
+        "/stac/bbox/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&assets=B01&rescale=0,1000&max_size=64&width=128&height=128"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
@@ -205,7 +205,7 @@ def test_part(httpx, rio, app):
     assert meta["height"] == 128
 
     response = app.get(
-        "/stac/crop/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&expression=B01_b1&rescale=0,1000&max_size=64"
+        "/stac/bbox/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&expression=B01_b1&rescale=0,1000&max_size=64"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
@@ -217,7 +217,7 @@ def test_part(httpx, rio, app):
 @patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_point(httpx, rio, app):
-    """test crop endpoints."""
+    """test point endpoints."""
     httpx.get = mock_RequestGet
     rio.open = mock_rasterio_open